This is an update that should make Rgraphviz work with Graphviz 2.10 and 2.12.
Essentially I have done the following
1) Made new preprocesser variables:
  GRAPHVIZ_2_2_TO_2_3
  GRAPHVIZ_2_4_TO_2_9
  GRAPHVIZ_2_10_TO_MORE
and removed the old ones, which were
  GRAPHVIZGT_1_16
  GRAPHVIZGT_2_4
The new ones have been incorporated into the
configure.in script as well as /src/Makevars.win and I have run autoconf
to redo the configure script included

2) Added a warning regarding compiling graphviz with optimization in the
README file

3) Added a new version check to layoutGraph essentially incorporating
the fact that version 2.8 is still broken regarding doing neato layouts
on graphs with singletons (the old check was clearly written before 2.8
was released). It is fixed in 2.10 though (and was strangely working in
2.5 and 2.7!)

4) Added a comment in the beginning of the Rgraphviz vignette saying
that a user may have problem running some of the examples in the
vignette if the graphviz version is 2.4, 2.6 or 2.8.

5) Added a sessionInfo() and graphvizVersion() call to the two vignettes

6) Fixed a deprecation in layingoutPathsways.Rnw (changed
hgu95av2LOCUSID to hguav2ENTREZID)

7) Made nummerous changes to the files in src. Essentially I have used
preprocessor macros to make the old code work as it has done all the
time, while other code is included in case graphviz has version >= 2.10.
These changes includes
  * A new header file needs to be included (common.h)
  * gvc->info[0] was changed to gvc->common.info[0] (graphvizVersion.c)
  * The label information in a Agnodeinfo_t, Agedgeinfo_t was changed, specifically the structure node->u.label->u.txt.line has a new name: node->u.label->u.txt.para (and similar for edge...) (doLayout.c)

8) I bumped the version number from 1.13.3 to 1.13.4

9) I also made a change to getDefaultAttrs changing the default node
name from "" to "\\N", which according to teh Graphviz documentation is
shorthand for nodename. The old stuff only worked in some special cases,
at lest in 2.10 and 2.12.



git-svn-id: https://hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/Rgraphviz@21433 bc3139a8-67e5-0310-9ffc-ced21a209358