Functionality to run a dot graph through graphviz
The commit adds c-code in src/dotToFile to run a
graph in dot-format through graphviz without going through
the Ragraph class.

The method in R is integrated into toFile, which has been
changed to a generic to support this. An example has been
added to the example section and the manual file updated and
renamed.